+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased] (1.2.0)
+### Changed
+- Upgrade to Rider 2023.3 EAP1 and later.
+
 ### Added
 - New setting to define previewer process working directory (see [#266](https://github.com/ForNeVeR/AvaloniaRider/pull/266) for details).